% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ref_kmers.R
\name{assign_by}
\alias{assign_by}
\alias{assign_by.cfdnalab_ref_kmer_frequencies}
\title{Return the reference k-mer window assignment rule.}
\usage{
assign_by(x, ...)

\method{assign_by}{cfdnalab_ref_kmer_frequencies}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab reference k-mer object.}

\item{...}{Reserved for future methods.}
}
\value{
A scalar character value.
}
\description{
Return the reference k-mer window assignment rule.
}
