% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ref_kmers.R
\name{canonical}
\alias{canonical}
\alias{canonical.cfdnalab_ref_kmer_frequencies}
\title{Return whether reference k-mers were canonicalized.}
\usage{
canonical(x, ...)

\method{canonical}{cfdnalab_ref_kmer_frequencies}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab reference k-mer object.}

\item{...}{Reserved for future methods.}
}
\value{
A scalar logical.
}
\description{
Return whether reference k-mers were canonicalized.
}
