% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R, R/lengths.R,
%   R/midpoints.R
\name{group_idx.cfdnalab_grouped_end_motif_counts}
\alias{group_idx.cfdnalab_grouped_end_motif_counts}
\alias{group_idx}
\alias{group_idx.cfdnalab_grouped_length_counts}
\alias{group_idx.cfdnalab_midpoint_profiles}
\title{Look up a group index.}
\usage{
\method{group_idx}{cfdnalab_grouped_end_motif_counts}(x, group_name, ...)

group_idx(x, ...)

\method{group_idx}{cfdnalab_grouped_length_counts}(x, group_name, ...)

\method{group_idx}{cfdnalab_midpoint_profiles}(x, group_name, ...)
}
\arguments{
\item{x}{A cfDNAlab object with group labels.}

\item{group_name}{Group name to look up.}

\item{...}{Method-specific lookup arguments.}
}
\value{
A scalar one-based integer group index.
}
\description{
Look up a group index.
}
