% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R, R/midpoints.R
\name{groups.cfdnalab_grouped_end_motif_counts}
\alias{groups.cfdnalab_grouped_end_motif_counts}
\alias{groups}
\alias{groups.cfdnalab_midpoint_profiles}
\title{Return group metadata.}
\usage{
\method{groups}{cfdnalab_grouped_end_motif_counts}(x, ...)

groups(x, ...)

\method{groups}{cfdnalab_midpoint_profiles}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab object with a group axis.}

\item{...}{Reserved for future methods.}
}
\value{
A data frame with one row per group.
}
\description{
Return group metadata.
}
