% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R, R/ref_kmers.R
\name{has_motif.cfdnalab_end_motif_counts}
\alias{has_motif.cfdnalab_end_motif_counts}
\alias{has_motif}
\alias{has_motif.cfdnalab_ref_kmer_frequencies}
\title{Test whether a motif label exists.}
\usage{
\method{has_motif}{cfdnalab_end_motif_counts}(x, motif, ...)

has_motif(x, ...)

\method{has_motif}{cfdnalab_ref_kmer_frequencies}(x, motif, ...)
}
\arguments{
\item{x}{A cfDNAlab object with motif labels.}

\item{motif}{Motif label to test.}

\item{...}{Method-specific lookup arguments.}
}
\value{
A scalar logical.
}
\description{
Test whether a motif label exists.
}
