% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ref_kmers.R
\name{kmer_size}
\alias{kmer_size}
\alias{kmer_size.cfdnalab_ref_kmer_frequencies}
\title{Return the reference k-mer size.}
\usage{
kmer_size(x, ...)

\method{kmer_size}{cfdnalab_ref_kmer_frequencies}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab reference k-mer object.}

\item{...}{Reserved for future methods.}
}
\value{
A scalar integer k-mer size.
}
\description{
Return the reference k-mer size.
}
