% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/lengths.R, R/midpoints.R
\name{length_bins}
\alias{length_bins}
\alias{length_bins.cfdnalab_length_counts}
\alias{length_bins.cfdnalab_midpoint_profiles}
\title{Return length-bin metadata.}
\usage{
length_bins(x, ...)

\method{length_bins}{cfdnalab_length_counts}(x, ...)

\method{length_bins}{cfdnalab_midpoint_profiles}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab object with length bins.}

\item{...}{Reserved for future methods.}
}
\value{
A data frame with one row per length bin.
}
\description{
Return length-bin metadata.
}
