% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/midpoints.R
\name{midpoint_array}
\alias{midpoint_array}
\alias{midpoint_array.cfdnalab_midpoint_profiles}
\title{Return the full midpoint count array.}
\usage{
midpoint_array(x, ...)

\method{midpoint_array}{cfdnalab_midpoint_profiles}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab midpoint-profile object.}

\item{...}{Reserved for future methods.}
}
\value{
A three-dimensional numeric array.
}
\description{
Return the full midpoint count array.
}
