% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ref_kmers.R
\name{motif_axis_kind}
\alias{motif_axis_kind}
\alias{motif_axis_kind.cfdnalab_ref_kmer_frequencies}
\title{Return the reference k-mer motif-axis kind.}
\usage{
motif_axis_kind(x, ...)

\method{motif_axis_kind}{cfdnalab_ref_kmer_frequencies}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab reference k-mer object.}

\item{...}{Reserved for future methods.}
}
\value{
A scalar character value, either \code{"motif"} or \code{"motif_group"}.
}
\description{
Return the reference k-mer motif-axis kind.
}
