% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R, R/ref_kmers.R
\name{motif_idx.cfdnalab_end_motif_counts}
\alias{motif_idx.cfdnalab_end_motif_counts}
\alias{motif_idx}
\alias{motif_idx.cfdnalab_ref_kmer_frequencies}
\title{Look up a motif index.}
\usage{
\method{motif_idx}{cfdnalab_end_motif_counts}(x, motif, ...)

motif_idx(x, ...)

\method{motif_idx}{cfdnalab_ref_kmer_frequencies}(x, motif, ...)
}
\arguments{
\item{x}{A cfDNAlab object with motif labels.}

\item{motif}{Motif label to look up.}

\item{...}{Method-specific lookup arguments.}
}
\value{
A scalar one-based integer motif index.
}
\description{
Look up a motif index.
}
