% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R
\name{motifs.cfdnalab_end_motif_counts}
\alias{motifs.cfdnalab_end_motif_counts}
\alias{motifs}
\title{Return end-motif metadata.}
\usage{
\method{motifs}{cfdnalab_end_motif_counts}(x, ...)

motifs(x, ...)
}
\arguments{
\item{x}{A cfDNAlab end-motif object.}

\item{...}{Reserved for future methods.}
}
\value{
A data frame with one row per motif.
}
\description{
Return end-motif metadata.
}
