% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R
\name{print.cfdnalab_end_motif_counts}
\alias{print.cfdnalab_end_motif_counts}
\title{Print an end-motif object.}
\usage{
\method{print}{cfdnalab_end_motif_counts}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab end-motif object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an end-motif object.
}
\keyword{internal}
