% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths.R
\name{print.cfdnalab_length_counts}
\alias{print.cfdnalab_length_counts}
\title{Print a length-count object.}
\usage{
\method{print}{cfdnalab_length_counts}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab length-count object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a length-count object.
}
\keyword{internal}
