% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midpoints.R
\name{print.cfdnalab_midpoint_profiles}
\alias{print.cfdnalab_midpoint_profiles}
\title{Print a midpoint-profile object.}
\usage{
\method{print}{cfdnalab_midpoint_profiles}(x, ...)
}
\arguments{
\item{x}{A \code{cfdnalab_midpoint_profiles} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a midpoint-profile object.
}
\keyword{internal}
