% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_kmers.R
\name{print.cfdnalab_ref_kmer_frequencies}
\alias{print.cfdnalab_ref_kmer_frequencies}
\title{Print a reference k-mer object.}
\usage{
\method{print}{cfdnalab_ref_kmer_frequencies}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab reference k-mer object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a reference k-mer object.
}
\keyword{internal}
