% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ref_kmers.R
\name{reference_contig_footprint}
\alias{reference_contig_footprint}
\alias{reference_contig_footprint.cfdnalab_ref_kmer_frequencies}
\title{Return the reference contig footprint.}
\usage{
reference_contig_footprint(x, ...)

\method{reference_contig_footprint}{cfdnalab_ref_kmer_frequencies}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab reference k-mer object.}

\item{...}{Reserved for future methods.}
}
\value{
JSON-decoded reference contig footprint metadata.
}
\description{
Return the reference contig footprint.
}
