% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R
\name{row_mode.cfdnalab_end_motif_counts}
\alias{row_mode.cfdnalab_end_motif_counts}
\alias{row_mode}
\title{Return the end-motif row mode.}
\usage{
\method{row_mode}{cfdnalab_end_motif_counts}(x, ...)

row_mode(x, ...)
}
\arguments{
\item{x}{A cfDNAlab end-motif object.}

\item{...}{Reserved for future methods.}
}
\value{
A scalar character value describing the row axis.
}
\description{
Return the end-motif row mode.
}
