% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/zarr_helpers.R
\name{schema_version}
\alias{schema_version}
\alias{schema_version.cfdnalab_zarr_store}
\title{Return the cfDNAlab schema version.}
\usage{
schema_version(x, ...)

\method{schema_version}{cfdnalab_zarr_store}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab loader object.}

\item{...}{Reserved for future methods.}
}
\value{
An integer schema version.
}
\description{
Return the cfDNAlab schema version.
}
