% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R
\name{storage_mode.cfdnalab_end_motif_counts}
\alias{storage_mode.cfdnalab_end_motif_counts}
\alias{storage_mode}
\title{Return the end-motif storage mode.}
\usage{
\method{storage_mode}{cfdnalab_end_motif_counts}(x, ...)

storage_mode(x, ...)
}
\arguments{
\item{x}{A cfDNAlab end-motif object.}

\item{...}{Reserved for future methods.}
}
\value{
A scalar character value, currently \code{"dense"} or \code{"sparse_coo"}.
}
\description{
Return the end-motif storage mode.
}
