% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R
\name{windows.cfdnalab_windowed_end_motif_counts}
\alias{windows.cfdnalab_windowed_end_motif_counts}
\alias{windows}
\title{Return window metadata.}
\usage{
\method{windows}{cfdnalab_windowed_end_motif_counts}(x, ...)

windows(x, ...)
}
\arguments{
\item{x}{A cfDNAlab object with window rows.}

\item{...}{Reserved for future methods.}
}
\value{
A data frame with one row per window.
}
\description{
Return window metadata.
}
