% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R, R/lengths.R,
%   R/midpoints.R
\name{group_metadata.cfdnalab_grouped_end_motif_counts}
\alias{group_metadata.cfdnalab_grouped_end_motif_counts}
\alias{group_metadata}
\alias{group_metadata.cfdnalab_grouped_length_counts}
\alias{group_metadata.cfdnalab_midpoint_profiles}
\title{Return group metadata.}
\usage{
\method{group_metadata}{cfdnalab_grouped_end_motif_counts}(x, ...)

group_metadata(x, ...)

\method{group_metadata}{cfdnalab_grouped_length_counts}(x, ...)

\method{group_metadata}{cfdnalab_midpoint_profiles}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab object with a group axis.}

\item{...}{Reserved for future methods.}
}
\value{
A data frame with one row per group.
}
\description{
Return group metadata.
}
