% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R
\name{has_motif.cfdnalab_end_motif_counts}
\alias{has_motif.cfdnalab_end_motif_counts}
\alias{has_motif}
\title{Test whether an end-motif label exists.}
\usage{
\method{has_motif}{cfdnalab_end_motif_counts}(x, motif, ...)

has_motif(x, ...)
}
\arguments{
\item{x}{A cfDNAlab end-motif object.}

\item{motif}{Motif label to test.}

\item{...}{Method-specific lookup arguments.}
}
\value{
A scalar logical.
}
\description{
Test whether an end-motif label exists.
}
