% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/lengths.R
\name{length_counts_vector}
\alias{length_counts_vector}
\alias{length_counts_vector.cfdnalab_global_length_counts}
\title{Return global length-count values as a vector.}
\usage{
length_counts_vector(x, ...)

\method{length_counts_vector}{cfdnalab_global_length_counts}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab global length-count object.}

\item{...}{Reserved for future methods.}
}
\value{
A named numeric vector with one value per length bin.
}
\description{
Return global length-count values as a vector.
}
