% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R, R/ref_kmers.R
\name{row_mode.cfdnalab_end_motif_counts}
\alias{row_mode.cfdnalab_end_motif_counts}
\alias{row_mode}
\alias{row_mode.cfdnalab_ref_kmer_frequencies}
\title{Return the output row mode.}
\usage{
\method{row_mode}{cfdnalab_end_motif_counts}(x, ...)

row_mode(x, ...)

\method{row_mode}{cfdnalab_ref_kmer_frequencies}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab object with matrix rows.}

\item{...}{Reserved for future methods.}
}
\value{
A scalar character value describing the row axis.
}
\description{
Return the output row mode.
}
