% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_motifs.R, R/generics.R, R/ref_kmers.R
\name{storage_mode.cfdnalab_end_motif_counts}
\alias{storage_mode.cfdnalab_end_motif_counts}
\alias{storage_mode}
\alias{storage_mode.cfdnalab_ref_kmer_frequencies}
\title{Return the output storage mode.}
\usage{
\method{storage_mode}{cfdnalab_end_motif_counts}(x, ...)

storage_mode(x, ...)

\method{storage_mode}{cfdnalab_ref_kmer_frequencies}(x, ...)
}
\arguments{
\item{x}{A cfDNAlab object with dense or sparse matrix output.}

\item{...}{Reserved for future methods.}
}
\value{
A scalar character value, currently \code{"dense"} or \code{"sparse_coo"}.
}
\description{
Return the output storage mode.
}
